+++ src/diff/oasDiff.ts
@@ -17,6 +17,7 @@ interface OasDiffFlags {
   format?: "json" | "console";
   dir?: boolean;
   "out-file"?: string;
+  "disable-normalize-directory-names"?: boolean;
 }
 
 /**
@@ -243,6 +244,28 @@ async function executeOasDiff(
 }
 
 /**
+ * Remove minor and patch versions from directory names and keep only major version
+ * Example: 'shopper-stores-oas-1.0.16' -> 'shopper-stores-oas-1'
+ * We want to keep the major version for multiple versions of the same API, ie: Shopper Baskets
+ *
+ * @param dirs - Array of directory objects with name and path properties
+ * @returns Copy of passed array with updated name property
+ */
+function normalizeDirectoryNames(
+  dirs: Array<{ name: string; path: string }>
+): Array<{ name: string; path: string }> {
+  return dirs.map((dir) => ({
+    ...dir,
+    // matches the pattern of the version number in the directory name, ie: -1.0.16
+    // extracts the major version number if available, otherwise the original match if no major version is found
+    name: dir.name.replace(/-\d+\.\d+\.\d+$/, (match) => {
+      const majorVersion = match.match(/^-\d+/)?.[0];
+      return majorVersion || match;
+    }),
+  }));
+}
+
+/**
  * Handle directory mode comparison logic
  *
  * @param baseApi - The base API directory
@@ -262,8 +285,13 @@ async function handleDirectoryMode(
   let hasErrors = false;
 
   // Find all exchange.json files and their parent directories
-  const baseExchangeDirs = await findExchangeDirectories(baseApi);
-  const newExchangeDirs = await findExchangeDirectories(newApi);
+  let baseExchangeDirs = await findExchangeDirectories(baseApi);
+  let newExchangeDirs = await findExchangeDirectories(newApi);
+
+  if (!flags["disable-normalize-directory-names"]) {
+    baseExchangeDirs = normalizeDirectoryNames(baseExchangeDirs);
+    newExchangeDirs = normalizeDirectoryNames(newExchangeDirs);
+  }
 
   const allDirNames = new Set([
     ...baseExchangeDirs.map((dir) => dir.name),